misc cleanup and refactor in health plugin
