+++ packages/health/lib/health.dart
@@ -12,4 +12,3 @@ part 'src/data_types.dart';
 part 'src/functions.dart';
 part 'src/health_data_point.dart';
 part 'src/health_factory.dart';
-part 'src/units.dart';
