--- packages/health/lib/src/units.dart
@@ -1,21 +1,0 @@
-part of health;
-
-/// List of all [HealthDataPoint] units.
-enum HealthDataUnit {
-  BEATS_PER_MINUTE,
-  CALORIES,
-  COUNT,
-  DEGREE_CELSIUS,
-  GRAMS,
-  KILOGRAMS,
-  METERS,
-  MILLIGRAM_PER_DECILITER,
-  MILLIMETER_OF_MERCURY,
-  MILLISECONDS,
-  MINUTES,
-  NO_UNIT,
-  PERCENTAGE,
-  SIEMENS,
-  UNKNOWN_UNIT,
-  LITER,
-}
